feat: implement Orca Security CSV parser
Extract shared helper functions into helpers.py to avoid circular
imports between parser.py and csv_parser.py. The CSV parser reads
Orca Security alert exports and maps fields to DefectDojo findings.

Authored by T. Walker - DefectDojo
